Cambio de Indice

+++ ArbolesDeHuffman/Huffman.cs
@@ -8,7 +8,7 @@ namespace ArbolesDeHuffman
 {
     public interface Huffman<T>
     {
-        int index { get; set; }
+        double index { get; set; }
         T SumaDeIndices(T a , T b);
     }
 }
